Move code from Ios to SslStream
