+++ openssl/src/ssl/mod.rs
@@ -1,16 +1,14 @@
 use libc::{c_int, c_void, c_long};
 use std::any::TypeId;
 use std::collections::HashMap;
-use std::ffi::{CStr, CString};
+use std::ffi::{CString};
 use std::fmt;
-use std::io;
 use std::io::prelude::*;
 use std::os::unix::io::RawFd;
 use std::mem;
-use std::net;
 use std::path::Path;
 use std::ptr;
-use std::sync::{Once, ONCE_INIT, Arc, Mutex};
+use std::sync::{Once, ONCE_INIT, Mutex};
 use std::ops::{Deref, DerefMut};
 use std::cmp;
 use std::any::Any;
@@ -21,11 +19,15 @@ use std::slice;
 
 use bio::{MemBio,SocketBio,Bio};
 use ffi;
-use ssl::error::{SslError, SslSessionClosed, StreamError, OpenSslErrors};
+use ssl::error::{SslError};
 use x509::{X509StoreContext, X509FileType, X509};
 use crypto::pkey::PKey;
 
 pub mod error;
+pub mod ssl_stream;
+
+pub use self::ssl_stream::{SslStream,StreamIo};
+
 #[cfg(test)]
 mod tests;
 
@@ -758,406 +760,4 @@ make_LibSslError! {
     ErrorZeroReturn = SSL_ERROR_ZERO_RETURN,
     ErrorWantConnect = SSL_ERROR_WANT_CONNECT,
     ErrorWantAccept = SSL_ERROR_WANT_ACCEPT
-}
-
-pub trait Io {
-    fn in_retry_wrapper<F>(&mut self, ssl: &Arc<Ssl>, mut blk: F)
-            -> Result<c_int, SslError> where F: FnMut(&Arc<Ssl>) -> c_int;
-}
-
-pub struct SocketIo;
-
-impl SocketIo {
-    pub fn new() -> SocketIo {
-        SocketIo 
-    }
-}
-
-impl Io for SocketIo {
-    fn in_retry_wrapper<F>(&mut self, ssl: &Arc<Ssl>, mut blk: F)
-            -> Result<c_int, SslError> where F: FnMut(&Arc<Ssl>) -> c_int
-    {
-        loop {
-            let ret = blk(ssl);
-            if ret > 0 {
-                return Ok(ret);
-            } else {
-                let e = ssl.get_error(ret);
-                match e {
-                    LibSslError::ErrorWantRead => return Ok(0),
-                    LibSslError::ErrorWantWrite => { try_ssl_stream!(self.flush(ssl)) }
-                    LibSslError::ErrorZeroReturn => return Err(SslSessionClosed),
-                    LibSslError::ErrorSsl => return Err(SslError::get()),
-                    err => panic!("unexpected error {:?} {:?}", err, io::Error::last_os_error()),
-                }
-            }
-        }
-    }
-}
-
-pub struct StreamIo<S:Read+Write> {
-    stream: S,
-    buf: Vec<u8>,
-}
-
-impl<S:Read+Write> StreamIo<S> {
-    pub fn new(stream: S) -> StreamIo<S> {
-        StreamIo {
-            stream: stream,
-            // Maximum TLS record size is 16k
-            // We're just using this as a buffer, so there's no reason to pay
-            // to memset it
-            buf: {
-                const CAP: usize = 16 * 1024;
-                let mut v = Vec::with_capacity(CAP);
-                unsafe { v.set_len(CAP); }
-                v
-            }
-        }
-    }
-
-    fn write_through(&mut self, ssl: &Arc<Ssl>) -> io::Result<()> {
-        io::copy(&mut *ssl.get_wbio::<MemBio>(), &mut self.stream).map(|_| ())
-    }
-}
-
-
-impl<S:Read+Write> Io for StreamIo<S> {
-    fn in_retry_wrapper<F>(&mut self, ssl: &Arc<Ssl>, mut blk: F)
-            -> Result<c_int, SslError> where F: FnMut(&Arc<Ssl>) -> c_int {
-        loop {
-            let ret = blk(ssl);
-            if ret > 0 {
-                return Ok(ret);
-            }
-
-            let e = ssl.get_error(ret);
-            match e {
-                LibSslError::ErrorWantRead => {
-                    try_ssl_stream!(self.flush(ssl));
-                    let len = try_ssl_stream!(self.stream.read(&mut self.buf[..]));
-                    if len == 0 {
-                        return Ok(0);
-                    }
-                    try_ssl_stream!(ssl.get_rbio::<MemBio>().write_all(&self.buf[..len]));
-                }
-                LibSslError::ErrorWantWrite => { try_ssl_stream!(self.flush(ssl)) }
-                LibSslError::ErrorZeroReturn => return Err(SslSessionClosed),
-                LibSslError::ErrorSsl => return Err(SslError::get()),
-                err => panic!("unexpected error {:?} {:?}", err, io::Error::last_os_error()),
-            }
-        }
-    }
-}
-
-/// A stream wrapper which handles SSL encryption for an underlying stream.
-#[derive(Clone)]
-pub struct SslStream<I:Io> {
-    io: I,
-    pub ssl: Arc<Ssl>,
-}
-
-impl SslStream<StreamIo<net::TcpStream>> {
-    /// Create a new independently owned handle to the underlying socket.
-    pub fn try_clone(&self) -> io::Result<SslStream<StreamIo<net::TcpStream>>> {
-        Ok(SslStream { 
-            io: StreamIo {
-                stream: try!(self.io.stream.try_clone()),
-                buf: self.io.buf.clone(),
-            },
-            ssl: self.ssl.clone(),
-        })
-    }
-}
-
-impl<S> fmt::Debug for SslStream<StreamIo<S>> where S: fmt::Debug+Read+Write {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        write!(fmt, "SslStream {{ stream: {:?}, ssl: {:?} }}", self.io.stream, self.ssl)
-    }
-}
-
-impl fmt::Debug for SslStream<SocketIo> {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        write!(fmt, "SslStream {{ ssl: {:?} }}", self.ssl)
-    }
-}
-
-impl<I:Io+SslRead> SslStream<I> {
-    fn new_base(ssl:Ssl, io: I) -> SslStream<I> {
-        SslStream {
-            io: io,
-            ssl: Arc::new(ssl),
-        }
-    }
-
-    /// Get the compression currently in use.  The result will be
-    /// either None, indicating no compression is in use, or a string
-    /// with the compression name.
-    pub fn get_compression(&self) -> Option<String> {
-        let ptr = unsafe { ffi::SSL_get_current_compression(self.ssl.ssl) };
-        if ptr == ptr::null() {
-            return None;
-        }
-
-        let meth = unsafe { ffi::SSL_COMP_get_name(ptr) };
-        let s = unsafe {
-            String::from_utf8(CStr::from_ptr(meth).to_bytes().to_vec()).unwrap()
-        };
-
-        Some(s)
-    }
-
-    /// Returns the protocol selected by performing Next Protocol Negotiation, if any.
-    ///
-    /// The protocol's name is returned is an opaque sequence of bytes. It is up to the client
-    /// to interpret it.
-    ///
-    /// This method needs the `npn` feature.
-    #[cfg(feature = "npn")]
-    pub fn get_selected_npn_protocol(&self) -> Option<&[u8]> {
-        self.ssl.get_selected_npn_protocol()
-    }
-
-    pub fn pending(&mut self) -> Result<usize, SslError> {
-        self.io.pending(&self.ssl)
-    }
-}
-
-impl SslStream<SocketIo> {
-    pub fn new_server_from_socket_from(ssl: Ssl) -> Result<SslStream<SocketIo>, SslError> {
-        let mut s = SslStream::new_base(ssl, SocketIo::new());
-        s.io.in_retry_wrapper(&s.ssl, |ssl| { ssl.accept() }).and(Ok(s))
-    }
-
-    /// Attempts to create a new SSL stream from a given `Ssl` instance.
-    pub fn new_from_socket_from(ssl: Ssl) -> Result<SslStream<SocketIo>, SslError> {
-        let mut s = SslStream::new_base(ssl, SocketIo::new());
-        s.io.in_retry_wrapper(&s.ssl, |ssl| { ssl.connect() }).and(Ok(s))
-    }
-
-    /// Creates a new SSL stream
-    pub fn new_from_socket(ctx: &SslContext, socket: RawFd) -> Result<SslStream<SocketIo>, SslError> {
-        let ssl = try!(Ssl::new_from_socket(ctx, socket));
-        SslStream::new_from_socket_from(ssl)
-    }
-
-    /// Creates a new SSL server stream
-    pub fn new_server_from_socket(ctx: &SslContext, socket: RawFd) -> Result<SslStream<SocketIo>, SslError> {
-        let ssl = try!(Ssl::new_from_socket(ctx, socket));
-        SslStream::new_server_from_socket_from(ssl)
-    }
-}
-
-impl<S:Read+Write> SslStream<StreamIo<S>> {
-    pub fn new_server_from(ssl: Ssl, stream: S) -> Result<SslStream<StreamIo<S>>, SslError> {
-        let mut s = SslStream::new_base(ssl, StreamIo::new(stream));
-        s.io.in_retry_wrapper(&s.ssl, |ssl| { ssl.accept() }).and(Ok(s))
-    }
-
-    /// Attempts to create a new SSL stream from a given `Ssl` instance.
-    pub fn new_from(ssl: Ssl, stream: S) -> Result<SslStream<StreamIo<S>>, SslError> {
-        let mut s = SslStream::new_base(ssl, StreamIo::new(stream));
-        s.io.in_retry_wrapper(&s.ssl, |ssl| { ssl.connect() }).and(Ok(s))
-    }
-
-    /// Creates a new SSL stream
-    pub fn new(ctx: &SslContext, stream: S) -> Result<SslStream<StreamIo<S>>, SslError> {
-        let ssl = try!(Ssl::new(ctx));
-        SslStream::new_from(ssl, stream)
-    }
-
-    /// Creates a new SSL server stream
-    pub fn new_server(ctx: &SslContext, stream: S) -> Result<SslStream<StreamIo<S>>, SslError> {
-        let ssl = try!(Ssl::new(ctx));
-        SslStream::new_server_from(ssl, stream)
-    }
-
-    /// Returns a mutable reference to the underlying stream.
-    ///
-    /// ## Warning
-    ///
-    /// `read`ing or `write`ing directly to the underlying stream will most
-    /// likely desynchronize the SSL session.
-    #[deprecated="use get_mut instead"]
-    pub fn get_inner(&mut self) -> &mut S {
-        self.get_mut()
-    }
-
-    /// Returns a reference to the underlying stream.
-    pub fn get_ref(&self) -> &S {
-        &self.io.stream
-    }
-
-    /// Returns a mutable reference to the underlying stream.
-    ///
-    /// ## Warning
-    ///
-    /// It is inadvisable to read from or write to the underlying stream as it
-    /// will most likely desynchronize the SSL session.
-    pub fn get_mut(&mut self) -> &mut S {
-        &mut self.io.stream
-    }
-}
-
-pub trait SslRead {
-    fn read(&mut self, ssl: &Arc<Ssl>, buf: &mut [u8]) -> io::Result<usize>;
-    fn pending(&mut self, ssl: &Arc<Ssl>) -> Result<usize, SslError>;
-}
-
-trait SslWrite {
-    fn write(&mut self, ssl: &Arc<Ssl>, buf: &[u8]) -> io::Result<usize>;
-    fn flush(&mut self, ssl: &Arc<Ssl>) -> io::Result<()>;
-}
-
-impl<I: Io+SslRead> Read for SslStream<I> {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        self.io.read(&self.ssl, buf)
-    }
-}
-
-impl SslRead for SocketIo {
-    fn read(&mut self, ssl: &Arc<Ssl>, buf: &mut [u8]) -> io::Result<usize> {
-        match self.in_retry_wrapper(&ssl, |ssl| { ssl.read(buf) }) {
-            Ok(len) => Ok(len as usize),
-            Err(SslSessionClosed) => Ok(0),
-            Err(StreamError(e)) => Err(e),
-            Err(e @ OpenSslErrors(_)) => {
-                Err(io::Error::new(io::ErrorKind::Other, e))
-            }
-        }
-    }
-
-    fn pending(&mut self, ssl: &Arc<Ssl>) -> Result<usize, SslError> {
-        let len = unsafe { ffi::SSL_pending(ssl.ssl) as usize };
-        Ok(len)
-    }
-}
-
-impl<S:Read+Write> SslRead for StreamIo<S> {
-    fn read(&mut self, ssl: &Arc<Ssl>, buf: &mut [u8]) -> io::Result<usize> {
-        match self.in_retry_wrapper(&ssl, |ssl| { ssl.read(buf) }) {
-            Ok(len) => Ok(len as usize),
-            Err(SslSessionClosed) => Ok(0),
-            Err(StreamError(e)) => Err(e),
-            Err(e @ OpenSslErrors(_)) => {
-                Err(io::Error::new(io::ErrorKind::Other, e))
-            }
-        }
-    }
-
-    fn pending(&mut self, ssl: &Arc<Ssl>) -> Result<usize, SslError> {
-        if ssl.get_rbio::<MemBio>().is_eof() {
-            try_ssl_stream!(self.flush(ssl));
-            let len = try_ssl_stream!(self.stream.read(&mut self.buf[..]));
-            try_ssl_stream!(ssl.get_rbio::<MemBio>().write_all(&self.buf[..len]));
-        }
-
-        let len = unsafe { ffi::SSL_pending(ssl.ssl) as usize };
-        Ok(len)
-    }
-}
-
-impl<I:Io+SslWrite> Write for SslStream<I> {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        self.io.write(&self.ssl, buf)
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        self.io.flush(&self.ssl)
-    }
-}
-
-impl<S:Read+Write> SslWrite for StreamIo<S> {
-    fn write(&mut self, ssl: &Arc<Ssl>, buf: &[u8]) -> io::Result<usize> {
-        match self.in_retry_wrapper(ssl, |ssl| ssl.write(buf)) {
-            Ok(len) => Ok(len as usize),
-            Err(SslSessionClosed) => Ok(0),
-            Err(StreamError(e)) => return Err(e),
-            Err(e @ OpenSslErrors(_)) => {
-                Err(io::Error::new(io::ErrorKind::Other, e))
-            }
-        }
-    }
-
-    fn flush(&mut self, ssl: &Arc<Ssl>) -> io::Result<()> {
-        try!(self.write_through(ssl));
-        self.stream.flush()
-    }
-}
-
-impl SslWrite for SocketIo {
-    fn write(&mut self, ssl: &Arc<Ssl>, buf: &[u8]) -> io::Result<usize> {
-        match self.in_retry_wrapper(ssl, |ssl| ssl.write(buf)) {
-            Ok(len) => Ok(len as usize),
-            Err(SslSessionClosed) => Ok(0),
-            Err(StreamError(e)) => return Err(e),
-            Err(e @ OpenSslErrors(_)) => {
-                Err(io::Error::new(io::ErrorKind::Other, e))
-            }
-        }
-    }
-
-    fn flush(&mut self, ssl: &Arc<Ssl>) -> io::Result<()> {
-        if ssl.get_rbio::<SocketBio>().flush() {
-            Ok(())
-        } else {
-            Err(io::Error::new(io::ErrorKind::Other, "SocketBio::flush() failed"))
-        }
-    }
-}
-
-/// A utility type to help in cases where the use of SSL is decided at runtime.
-#[derive(Debug)]
-pub enum MaybeSslStream<S> where S: Read+Write {
-    /// A connection using SSL
-    Ssl(SslStream<StreamIo<S>>),
-    /// A connection not using SSL
-    Normal(S),
-}
-
-impl<S> Read for MaybeSslStream<S> where S: Read+Write {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        match *self {
-            MaybeSslStream::Ssl(ref mut s) => s.read(buf),
-            MaybeSslStream::Normal(ref mut s) => s.read(buf),
-        }
-    }
-}
-
-impl<S> Write for MaybeSslStream<S> where S: Read+Write {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        match *self {
-            MaybeSslStream::Ssl(ref mut s) => s.write(buf),
-            MaybeSslStream::Normal(ref mut s) => s.write(buf),
-        }
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        match *self {
-            MaybeSslStream::Ssl(ref mut s) => s.flush(),
-            MaybeSslStream::Normal(ref mut s) => s.flush(),
-        }
-    }
-}
-
-impl<S> MaybeSslStream<S> where S: Read+Write {
-    /// Returns a reference to the underlying stream.
-    pub fn get_ref(&self) -> &S {
-        match *self {
-            MaybeSslStream::Ssl(ref s) => s.get_ref(),
-            MaybeSslStream::Normal(ref s) => s,
-        }
-    }
-
-    /// Returns a mutable reference to the underlying stream.
-    ///
-    /// ## Warning
-    ///
-    /// It is inadvisable to read from or write to the underlying stream.
-    pub fn get_mut(&mut self) -> &mut S {
-        match *self {
-            MaybeSslStream::Ssl(ref mut s) => s.get_mut(),
-            MaybeSslStream::Normal(ref mut s) => s,
-        }
-    }
 }
